# api/openapi-spec.yaml
openapi: 3.0.3
info:
  title: Flowfish Platform API
  description: |
    # Flowfish Platform API
    
    REST API documentation for the eBPF-based Kubernetes/OpenShift application communication and dependency mapping platform.
    
    ## Key Features
    
    - **Automatic Discovery** — Automatically detect inter-pod communications using eBPF
    - **Real-Time Dependency Map** — Live visualization of service dependency graphs
    - **Anomaly Detection** — Intelligent anomaly detection powered by statistical analysis
    - **Multi-Cluster** — Manage multiple clusters from a single interface
    - **Integration** — Dependency intelligence APIs for CI/CD pipeline integration
    - **Pod & Deployment Annotations** — Rich metadata enrichment including annotations merged from owning Deployments/StatefulSets
    - **Import/Export** — Data import and export in multiple formats
    
    ## Authentication
    
    The API supports two authentication methods:
    
    1. **JWT (JSON Web Token)** — Send a POST request to `/api/v1/auth/login` with username and password, then include the returned `access_token` in subsequent requests via `Authorization: Bearer <token>` header.
    2. **API Key** — Generate an API key from Settings > API Keys, then include it via `X-API-Key: <key>` header. Recommended for CI/CD pipelines and AI agent integrations.
    
    ## Rate Limiting
    
    - **Authenticated Users**: 1,000 requests/hour
    - **Unauthenticated**: 100 requests/hour
    
...
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authorization header using the Bearer scheme
...
security:
  - BearerAuth: []
paths:
...
  /events/histogram:
    get:
      tags:
        - Events
      summary: Event time histogram
      description: Returns event counts aggregated into time buckets for the given range and filters
      parameters:
        - name: cluster_id
          in: query
          required: true
          schema:
            type: integer
        - name: analysis_id
          in: query
          schema:
            type: integer
        - name: event_types
          in: query
          schema:
            type: string
          description: Comma-separated event type identifiers
        - name: start_time
          in: query
          schema:
            type: string
            format: date-time
        - name: end_time
          in: query
          schema:
            type: string
            format: date-time
        - name: bucket_count
          in: query
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Histogram buckets and metadata
          content:
            application/json:
              schema:
                type: object
                properties:
                  buckets:
                    type: array
                    items:
                      type: object
                      additionalProperties: true
                  time_range:
                    type: object
                    additionalProperties: true
                  interval_seconds:
                    type: number
                  total_events:
                    type: integer
...
  /l7/events/http:
    get:
      tags: [L7 Events]
      summary: Query L7 HTTP flow events
      parameters:
        - name: analysis_id
          in: query
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: method
          in: query
          schema: { type: string }
        - name: path
          in: query
          schema: { type: string }
        - name: status_code
          in: query
          schema: { type: integer }
        - name: start_time
          in: query
          schema: { type: string, format: date-time }
        - name: end_time
          in: query
          schema: { type: string, format: date-time }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Paginated HTTP flow events
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/events/grpc:
    get:
      tags: [L7 Events]
      summary: Query L7 gRPC flow events
      parameters:
        - name: analysis_id
          in: query
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: grpc_service
          in: query
          schema: { type: string }
        - name: grpc_method
          in: query
          schema: { type: string }
        - name: grpc_status_code
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Paginated gRPC flow events
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/events/dns:
    get:
      tags: [L7 Events]
      summary: Query L7 DNS flow events
      parameters:
        - name: analysis_id
          in: query
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: query_name
          in: query
          schema: { type: string }
        - name: query_type
          in: query
          schema: { type: string }
        - name: response_code
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 100 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Paginated DNS flow events
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/events/stats:
    get:
      tags: [L7 Events]
      summary: Aggregate L7 event statistics across protocols
      parameters:
        - name: analysis_id
          in: query
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: start_time
          in: query
          schema: { type: string, format: date-time }
        - name: end_time
          in: query
          schema: { type: string, format: date-time }
      responses:
        '200':
          description: Per-protocol event counts and totals
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
  /l7/events/histogram:
    get:
      tags: [L7 Events]
      summary: L7 HTTP request histogram (5-min buckets)
      parameters:
        - name: analysis_id
          in: query
          schema: { type: integer }
        - name: cluster_id
          in: query
          schema: { type: integer }
        - name: namespace
          in: query
          schema: { type: string }
        - name: start_time
          in: query
          schema: { type: string, format: date-time }
        - name: end_time
          in: query
          schema: { type: string, format: date-time }
        - name: bucket_count
          in: query
          schema: { type: integer, default: 60, minimum: 1, maximum: 200 }
      responses:
        '200':
          description: Time-bucketed request counts with error rates
          content:
            application/json:
              schema: { type: object, additionalProperties: true }